fix: authentication error 수정

+++ src/main/kotlin/com/b05/newsfeedproject/domain/exception/ModelNotFoundException.kt
@@ -1,6 +1,6 @@
 package com.b05.newsfeedproject.domain.exception
 
-data class ModelNotFoundException(val modelName:String,val id:Long):RuntimeException(
+data class ModelNotFoundException(val modelName:String,val id:Int):RuntimeException(
             "${modelName} is not found in id:${id}"
 )
