+++ src/main/kotlin/com/b05/newsfeedproject/domain/user/dto/UserResponse.kt
@@ -3,7 +3,7 @@ package com.b05.newsfeedproject.domain.user.dto
 import java.util.*
 
 data class UserResponse(
-        var id: Long?,
+        var id: Int?,
         var email: String,
         var nickName: String,
         var password:String,
