+++ src/main/kotlin/com/b05/newsfeedproject/domain/user/service/UserService.kt
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestBody
 
 interface UserService {
-    fun getUserById(userId: Long):UserResponse
+    fun getUserById(userId: Int):UserResponse
     fun signup(request: SignupUserRequest): UserResponse
-    fun updateUser(userId: Long, request: UpdateUserRequest): UserResponse
+    fun updateUser(userId: Int, request: UpdateUserRequest): UserResponse
     fun signIn(request: SignInUserRequest): SignInResponse
     fun signout()
 }
